requirements.txt
Pillow==9.4.0
pandas==2.0.3